x.stream - format imports & gofmt
