--- x/stream/keeper/events.go
@@ -1,1 +1,0 @@
-package keeper
