+++ x/stream/keeper/msg_server.go
@@ -2,10 +2,12 @@ package keeper
 
 import (
 	"context"
+
 	"github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+
 	"github.com/unification-com/mainchain/x/stream/types"
 )
