+++ x/stream/keeper/query.go
@@ -2,10 +2,12 @@ package keeper
 
 import (
 	"context"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/unification-com/mainchain/x/stream/types"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+
+	"github.com/unification-com/mainchain/x/stream/types"
 )
 
 var _ types.QueryServer = Keeper{}
