+++ x/stream/module_simulation.go
@@ -5,6 +5,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
+
 	"github.com/unification-com/mainchain/x/stream/simulation"
 	"github.com/unification-com/mainchain/x/stream/types"
 )
@@ -14,17 +15,17 @@ var (
 	_ = baseapp.Paramspace
 )
 
-// GenerateGenesisState creates a randomized GenState of the feegrant module.
+// GenerateGenesisState creates a randomized GenState of the stream module.
 func (AppModule) GenerateGenesisState(simState *module.SimulationState) {
 	simulation.RandomizedGenState(simState)
 }
 
-// RegisterStoreDecoder registers a decoder for feegrant module's types
+// RegisterStoreDecoder registers a decoder for stream module's types
 func (am AppModule) RegisterStoreDecoder(sdr sdk.StoreDecoderRegistry) {
 	sdr[types.StoreKey] = simulation.NewDecodeStore(am.cdc)
 }
 
-// WeightedOperations returns all the feegrant module operations with their respective weights.
+// WeightedOperations returns all the stream module operations with their respective weights.
 func (am AppModule) WeightedOperations(simState module.SimulationState) []simtypes.WeightedOperation {
 	return simulation.WeightedOperations(
 		simState.AppParams, simState.Cdc,
