+++ x/stream/simulation/decoder.go
@@ -3,8 +3,10 @@ package simulation
 import (
 	"bytes"
 	"fmt"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/types/kv"
+
 	"github.com/unification-com/mainchain/x/stream/types"
 )
