+++ x/stream/simulation/proposals.go
@@ -1,12 +1,14 @@
 package simulation
 
 import (
+	"math/rand"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/address"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
+
 	"github.com/unification-com/mainchain/x/stream/types"
-	"math/rand"
 )
 
 // Simulation operation weights constants
