+++ x/stream/types/utils_test.go
@@ -1,11 +1,13 @@
 package types_test
 
 import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stretchr/testify/require"
-	"github.com/unification-com/mainchain/x/stream/types"
 	"testing"
 	"time"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/stretchr/testify/require"
+
+	"github.com/unification-com/mainchain/x/stream/types"
 )
 
 func TestPeriodEnumFromString(t *testing.T) {
